Implemented a responsive landing page
An interactive navigation menu that changes style when hovering over a menu item. The navigation menu has a fixed position and is visible on all pages through a circular button on the top left of the screen that drops the menu down.
I also used GSAP for fluid page animations to provide a relaxing and seamless user experience.